Moving CreateBatchOptions to correct location.
