+++ sdk/servicebus/azure-messaging-servicebus/src/samples/java/com/azure/messaging/servicebus/ServiceBusSenderAsyncClientJavaDocCodeSamples.java
@@ -4,7 +4,7 @@
 package com.azure.messaging.servicebus;
 
 import com.azure.identity.DefaultAzureCredentialBuilder;
-import com.azure.messaging.servicebus.administration.models.CreateBatchOptions;
+import com.azure.messaging.servicebus.models.CreateBatchOptions;
 import reactor.core.Exceptions;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
